added health system and lives to screen
