+++ src/Level/Player.java
@@ -3,12 +3,15 @@ package Level;
 import Engine.Key;
 import Engine.KeyLocker;
 import Engine.Keyboard;
+import Game.ScreenCoordinator;
 import GameObject.GameObject;
 import GameObject.SpriteSheet;
 import Utils.AirGroundState;
 import Utils.Direction;
+import Utils.Stopwatch;
 
 import java.util.ArrayList;
+import java.util.Timer;
 
 public abstract class Player extends GameObject {
     //Score
@@ -20,6 +23,13 @@ public abstract class Player extends GameObject {
         return score;
     }
     //END SCORE
+
+    //LIVES
+    private int life = 3;
+    public void loseLife() {life--;}
+    public int getLife() {return life;}
+
+    //END LIVES
 
 
     // values that affect player movement
@@ -313,12 +323,26 @@ public abstract class Player extends GameObject {
         }
     }
 
+
+
     // other entities can call this method to hurt the player
     public void hurtPlayer(MapEntity mapEntity) {
         if (!isInvincible) {
             // if map entity is an enemy, kill player on touch
             if (mapEntity instanceof Enemy) {
-                levelState = LevelState.PLAYER_DEAD;
+
+                setPlayerState(PlayerState.JUMPING);
+                life--;
+
+                System.out.println("Lives Left" + life);
+                //System.out.println(life);
+                if(life <= 0) {
+                    levelState = LevelState.PLAYER_DEAD;
+                }else{
+
+                }
+
+
             }
         }
     }
@@ -364,7 +388,7 @@ public abstract class Player extends GameObject {
         }
         // if death animation not on last frame yet, continue to play out death animation
         else if (currentFrameIndex != getCurrentAnimation().length - 1) {
-          super.update();
+            super.update();
         }
         // if death animation on last frame (it is set up not to loop back to start), player should continually fall until it goes off screen
         else if (currentFrameIndex == getCurrentAnimation().length - 1) {
@@ -407,3 +431,4 @@ public abstract class Player extends GameObject {
         listeners.add(listener);
     }
 }
+
